[refactor]  월별, 일별 방문자 api 통일
- 월별, 일별 방문자 api 통일
  query에 type을 `daily | monthly`로 받아서 분기 처리하도록 구현
